% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ext_resources.R
\name{add_ext_resources}
\alias{add_ext_resources}
\title{Add external resources to the shiny app}
\usage{
add_ext_resources(data_dir)
}
\arguments{
\item{data_dir}{The (server-side) file-path for the directory containing the data for
presentation. This should contain a \code{scanner_output} subdirectory, which will be mounted as
/data/scanner_output/ in the browser.}
}
\description{
Function is internally used to add external resources inside the Shiny application.
}
