% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_clusters}
\alias{all_clusters}
\title{Data set containing the vector of all clusters}
\format{
A character vector of cluster IDs
}
\usage{
all_clusters
}
\description{
Data set containing the vector of all clusters
}
\keyword{datasets}
