% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_sequences}
\alias{available_sequences}
\title{function to return sequence options}
\usage{
available_sequences(data_dir)
}
\arguments{
\item{data_dir}{The data directory for the app. Must contain a "sequences" subdirectory.}
}
\description{
function to return sequence options
}
