% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{create_girafe}
\alias{create_girafe}
\title{Convert a ggplot object into an interactive {ggiraph} object}
\usage{
create_girafe(
  ggobj,
  widget_choice,
  width_svg,
  height_svg,
  suppress_warnings = FALSE
)
}
\arguments{
\item{ggobj}{The ggplot2/ggtree object.}

\item{widget_choice}{Scalar character. Describes the type of plot that is contained in
\code{ggobj}. Typically a file basename that includes the file extension, of the form "tree-XXX.rds"
(for \code{{ggtree}} objects) or "sina-XXX.rds" (for \code{{ggplot2}} scatter plots).}

\item{width_svg, height_svg}{Scalar numeric. The width/height of the output plot.}

\item{suppress_warnings}{Scalar logical. Should warnings from \code{ggiraph::girafe()} be printed
to the console?}
}
\description{
Convert a ggplot object into an interactive {ggiraph} object
}
