% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{create_sequences_lookup}
\alias{create_sequences_lookup}
\title{Function to save a CSV file of all sequences for all clusterIDs}
\usage{
create_sequences_lookup(data_dir)
}
\arguments{
\item{data_dir}{The data directory for the application. Must have a \code{scanner_output}
subdirectory.}
}
\description{
Function to save a CSV file of all sequences for all clusterIDs
}
