% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_panel}
\alias{display_panel}
\title{function to create UI for the Well panel in the tabs}
\usage{
display_panel(to_display)
}
\arguments{
\item{to_display}{UI element to be displayed in the well panel}
}
\description{
function to create UI for the Well panel in the tabs
}
