% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloader_tab_panel}
\alias{downloader_tab_panel}
\title{function to create display panel inside tabset panel witin module}
\usage{
downloader_tab_panel(title, chooser_id, download_button_id, panel)
}
\arguments{
\item{title}{Scalar string defining title of tab panel}

\item{chooser_id}{Scalar string of ID of the dropdown menu to display to choose files}

\item{download_button_id}{Scalar string of ID of the download button to download files}

\item{panel}{UI element to display on the right hand side of the panel}
}
\description{
function to create display panel inside tabset panel witin module
}
