% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_filetype}
\alias{filter_by_filetype}
\title{function to filter files in each data folder based on file extension}
\usage{
filter_by_filetype(filenames, filetypes)
}
\arguments{
\item{filenames}{Vector of all file names in the folder}

\item{filetypes}{Vector of which filetypes to filter by}
}
\description{
function to filter files in each data folder based on file extension
}
