% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_clusters}
\alias{get_all_clusters}
\title{function to return list of all clusters from folder name}
\usage{
get_all_clusters(filename = "inst/www/data/wcdemo/scanner_output")
}
\arguments{
\item{filename}{Character string for
filename of folder containing all outputs}
}
\description{
function to return list of all clusters from folder name
}
