% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Obtain the optionally-specified data directory for the app}
\usage{
get_data_dir()
}
\value{
Scalar string. The data-directory for use in the app.
}
\description{
The user can specify a data-directory by specifying the environment variable \code{APP_DATA_DIR}.
When given, this directory must have subdirectories: \code{mutations}, \code{scanner_output}, \code{sequences},
\code{treeview}.
When not specified, data will be taken from the package-embedded directory \verb{/app/www/data}.
}
