% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filename}
\alias{get_filename}
\title{function to return folder name}
\usage{
get_filename(type)
}
\arguments{
\item{type}{Character string detailing type of widget to show}
}
\description{
function to return folder name
}
