% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_mutations}
\alias{get_unique_mutations}
\title{function to return list of all mutations}
\usage{
get_unique_mutations(filename)
}
\arguments{
\item{filename}{Character string for filename of csv containing mutations}
}
\description{
function to return list of all mutations
}
