% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reformat_table}
\alias{reformat_table}
\title{function to tidy up table output}
\usage{
reformat_table(table_to_display)
}
\arguments{
\item{table_to_display}{Data frame or tibble containing messy outputs}
}
\description{
function to tidy up table output
}
