% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selected_seq_nodes}
\alias{selected_seq_nodes}
\title{Which nodes have a given sequence}
\usage{
selected_seq_nodes(chosen_sequence)
}
\arguments{
\item{chosen_sequence}{String for the user selected sequence}
}
\description{
Which nodes have a given sequence
}
