% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_treeview.R
\name{treeviewServer}
\alias{treeviewServer}
\title{Shiny server function for presenting the treeview data}
\usage{
treeviewServer(id)
}
\arguments{
\item{id}{Shiny/HTML ID for the module.}
}
\value{
Creates a \code{moduleServer} that returns a reactive containing the cluster ID for
whichever node the user most recently hovered on.
}
\description{
Shiny server function for presenting the treeview data
}
