% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_treeview.R
\name{treeviewUI}
\alias{treeviewUI}
\title{Shiny module for presenting the treeview data}
\usage{
treeviewUI(id)
}
\arguments{
\item{id}{Shiny/HTML ID for the module.}
}
\value{
fluidRow UI elements for the module.
}
\description{
Shiny module for presenting the treeview data
}
