% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{create_all_node_lookups}
\alias{create_all_node_lookups}
\title{function to create lookups for nodes for all treeviews}
\usage{
create_all_node_lookups()
}
\description{
function to create lookups for nodes for all treeviews
}
