% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{create_node_lookup}
\alias{create_node_lookup}
\title{function to create lookup for a single treeview}
\usage{
create_node_lookup(widgetChoice)
}
\description{
function to create lookup for a single treeview
}
