% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_files.R
\name{get_all_files}
\alias{get_all_files}
\title{Get all file names in a folder of the data relating to a single cluster}
\usage{
get_all_files(cluster_choice)
}
\arguments{
\item{cluster_choice}{character relating to a folder name in inst/data}
}
\description{
Get all file names in a folder of the data relating to a single cluster
}
