% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filename}
\alias{get_filename}
\title{Obtain the file-path for a tree-view}
\usage{
get_filename(type, data_dir)
}
\arguments{
\item{type}{Choice of tree-view widget to show}

\item{data_dir}{The data directory for the app. Must contain a "treeview" subdirectory.}
}
\description{
Obtain the file-path for a tree-view
}
