% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{process_seq_table}
\alias{process_seq_table}
\title{function to get lookup table of clusterID and sequence}
\usage{
process_seq_table(selected_folder)
}
\arguments{
\item{selected_folder}{Folder name relating to a single clusterID}
}
\description{
function to get lookup table of clusterID and sequence
}
