% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{update_data}
\alias{update_data}
\title{function to be run anytime the data is updated
wrapper around other required functions
ideally these will be added to {tfpscanner} in the longer term
and the outputs transferred instead}
\usage{
update_data(treeview = "tree-logistic_growth_rate.rds")
}
\arguments{
\item{treeview}{RDS file containing an existing treeview plot}
}
\description{
function to be run anytime the data is updated
wrapper around other required functions
ideally these will be added to {tfpscanner} in the longer term
and the outputs transferred instead
}
